Refactor : 엔티티 기본키 Long -> Integer

See merge request s10-webmobile2-sub2/S10P12A610!29

+++ backend/src/main/java/com/ssafy/togeball/domain/chatroom/repository/ChatroomRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ChatroomRepository extends JpaRepository<Chatroom, Long>, CustomChatroomRepository {
+public interface ChatroomRepository extends JpaRepository<Chatroom, Integer>, CustomChatroomRepository {
 
     List<Chatroom> findAllByTitleContaining(String title);
 }
